requirements.txt
django-environ==0.6.0
Django==3.2.7
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
drf-jwt==1.19.1
django-versatileimagefield==2.2
gunicorn==20.1.0
whitenoise==5.3.0
psycopg2-binary==2.9.1
django-heroku
